% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_api_blastn.R
\name{blast_api_blastn}
\alias{blast_api_blastn}
\title{Wrapper for blast api blastn}
\usage{
blast_api_blastn(seq)
}
\arguments{
\item{seq}{a sequence in character format}
}
\value{
an info vector with RID and RTOE (estimate time in second)
}
\description{
use for blastn the nucleotide sequence on ncbi blast database
notes: only get 20 result, slower then megablast
submit when there are no hits
}
