% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_api_megablast.R
\name{blast_api_megablast}
\alias{blast_api_megablast}
\title{Wrapper for blas api megablast}
\usage{
blast_api_megablast(seq)
}
\arguments{
\item{seq}{a sequence in character format}
}
\value{
an info vector with RID and RTOE (estimate time in second)
}
\description{
use to start the blastn program for nucleotide sequence
notes: at the start of the wrapper for fast result
}
