% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_api_wrapper.R
\name{blast_api_wrapper}
\alias{blast_api_wrapper}
\title{Wrapper for blast api}
\usage{
blast_api_wrapper(seq, verbose = FALSE)
}
\arguments{
\item{seq}{a sequence in character format}
}
\value{
a GET result that can be parse with read_html
}
\description{
use for blast the nucleotide sequence on ncbi blast database
notes: only get 20 result, run with megablast option for fast query
}
\examples{
data(rbcLseq)
res = blast_api_wrapper(rbcLseq)
data("rbcLres")
res |>
 extract_table_fromGET()
}
