% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table_fromGET.R
\name{extract_table_fromGET}
\alias{extract_table_fromGET}
\title{Extract table from api request}
\usage{
extract_table_fromGET(res)
}
\arguments{
\item{res}{the GET result}
}
\value{
data frame
}
\description{
Extract table from api request
}
