% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiBlastR-package.R
\docType{package}
\name{ncbiBlastR-package}
\alias{ncbiBlastR}
\alias{ncbiBlastR-package}
\title{ncbiBlastR: send the sequence to NCBI blast server via API request}
\description{
Send DNA sequence to blastn with megablast mode on. database nt/nr If fail to get hits, submit again with blastn normal mode
}
\keyword{internal}
