% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html_comment.R
\name{parse_html_comment}
\alias{parse_html_comment}
\title{Parse html comment from the query request}
\usage{
parse_html_comment(res)
}
\arguments{
\item{res}{response object from GET in html format}
}
\value{
a named vector of strings
}
\description{
This function will parse all the comment, then search for pattern 'QBlastInfo'
}
