% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_result.R
\name{query_result}
\alias{query_result}
\title{Query the result of RID from ncbi blast}
\usage{
query_result(rid)
}
\arguments{
\item{rid}{unique RID from ncbi blast PUT command}
}
\value{
response in html format
}
\description{
use httr::GET
}
