% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_searchinfo.R
\name{query_searchinfo}
\alias{query_searchinfo}
\title{Query the search info of RID from ncbi blast}
\usage{
query_searchinfo(rid)
}
\arguments{
\item{rid}{unique RID from ncbi blast PUT command}
}
\value{
response in html format
}
\description{
use httr::GET
}
