% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbcLres}
\alias{rbcLres}
\title{blast result for the rbcL gene sequence}
\format{
## `rbcLres`
an html response that can be parsed with xml2 library
\describe{
  \item{the goal is to extract table data from this html response}
  \item{if the response has no table with data, send error message}
}
}
\source{
<blastn api>
}
\usage{
rbcLres
}
\description{
the result from blast_api_wrapper function. It is run with megablast method.
The object has class "response" from httr::GET function
}
\keyword{datasets}
