% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rbcLseq}
\alias{rbcLseq}
\title{DNA sequence for the rbcL gene}
\format{
A string with 1392 characters:
\itemize{
  \item a sequence of ACTG character
  \item read from fasta file
  \item header: ENA|D28335|D28335.1 Clintonia uniflora chloroplast rbcL gene
}
The raw data can be found in examples code
}
\source{
<https://www.ebi.ac.uk/ena/browser/view/D28335>
}
\usage{
rbcLseq
}
\description{
the Sequence is downloaded as a fasta file from ENA website
}
\examples{
path = system.file("extdata", package="ncbiBlastR")
files = list.files(path, full.names=TRUE)
readLines(files[[1]])
}
\keyword{datasets}
