% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_and_check.R
\name{wait_and_check}
\alias{wait_and_check}
\title{wait and check if there are hits}
\usage{
wait_and_check(info)
}
\arguments{
\item{info}{parsed from GET result}
}
\description{
Use after submit of
}
\keyword{internal}
